Fix how updated_by gets assigned (#14678)
Resolves #14205

### Description

  * Create concern to standardize how `updated_by` gets set

### Acceptance Criteria
- [x]  save the current user whenever an update is done
- [x] use before_save instead of before_update to set the value, so it'll be set on create as well
- [x] don't save a nil value if RequestStore[:current_user] isn't set (see the implementation in the VirtualHearing model)
- [x] verify that this change won't invalidate assumptions we're making as we use updated_by elsewhere in the code (for example, in the Hearing model, we're using the timezone of the updated_by user to convert hearing times)
