+++ app/models/legacy_hearing.rb
@@ -8,6 +8,7 @@ class LegacyHearing < CaseflowRecord
   include HasVirtualHearing
   include HearingLocationConcern
   include HearingTimeConcern
+  include UpdatedByUserConcern
 
   # When these instance variable getters are called, first check if we've
   # fetched the values from VACOLS. If not, first fetch all values and save them
@@ -28,7 +29,6 @@ class LegacyHearing < CaseflowRecord
   belongs_to :appeal, class_name: "LegacyAppeal"
   belongs_to :user # the judge
   belongs_to :created_by, class_name: "User"
-  belongs_to :updated_by, class_name: "User"
   has_many :hearing_views, as: :hearing
   has_many :appeal_stream_snapshots, foreign_key: :hearing_id
   has_one :hearing_location, as: :hearing
@@ -52,7 +52,6 @@ class LegacyHearing < CaseflowRecord
   delegate :timezone, :name, to: :regional_office, prefix: true
 
   before_create :assign_created_by_user
-  before_update :assign_updated_by_user
 
   CO_HEARING = "Central"
   VIDEO_HEARING = "Video"
@@ -323,9 +322,5 @@ class LegacyHearing < CaseflowRecord
 
   def assign_created_by_user
     self.created_by ||= RequestStore[:current_user]
-  end
-
-  def assign_updated_by_user
-    self.updated_by ||= RequestStore[:current_user]
   end
 end
